+++ _src/interfaces/eventConstructor.ts
@@ -1,0 +1,4 @@
+import { ConstructorType } from '../objects';
+
+export type EventConstructor<T = any> = ConstructorType<T> & { eventName: string };
+export type EventNameOrConstructor<T = any> = EventConstructor<T> | string;