+++ _src/tests/interceptors/client/test.send-event-subscription.interceptor.ts
@@ -1,11 +1,12 @@
-import { EventSubscriptionSendInterceptor } from '../../../interfaces/eventSubscriptionInterceptor';
 import { EventSubscription } from '../../../interfaces/eventSubscription';
 import { TestSubscriptionData } from '../../events/eventBase';
 import { session } from '../../session';
+import { EventSubscriptionSendInterceptor } from '../../../interfaces/eventSubscriptionInterceptor';
+import { EventNameOrConstructor } from '../../../interfaces/eventConstructor';
 
 export class TestEventSubscriptionSendInterceptor implements EventSubscriptionSendInterceptor<any, TestSubscriptionData> {
 
-	async execute(eventClass: new (data?: any) => any, subscription: EventSubscription<TestSubscriptionData>): Promise<void> {
+	async execute(eventNameOrConstructor: EventNameOrConstructor, subscription: EventSubscription<TestSubscriptionData>): Promise<void> {
 		subscription.additionalData = {
 			token: session.token,
 			tenant: session.tenant